Fixed failing test_generate_models_from_stem
This failure exposed an error in which self.parent.count in KmerNode.populate_sisters was treated as standing in for sum of children of that parent.

I added the script bin/summarize_children_of_prefix.py to help debug
this issue, it may come in handly in the future.

also changed imports for an unrelated issue where pytest wanted
different imports from everything else.